Refactor to support both deep/shallow models
